+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.15.2] - Jul 17, 2023
+
+### Changed
+
+- Move the small loader to a global component
+- Update the similar books section to have a more consistent design
+- Update README text
+
+### Fixed
+
+- Fix the similar books section
+
 ## [2.15.1] - Jul 16, 2023
 
 ### Added